requirements.txt
Flask==2.3.2
pyttsx3==2.90
wikipedia-api==0.5.8
pywhatkit==5.4
pyautogui==0.9.54
gunicorn
